feat(web): Grant - Mobile view and sidemenu refactor (#17214)
* WIP action card & alert banner for grant

* feat: add text from cms

* fix: UI tweas and starting mobile menu

* feat: mobile view

* feat: add disabled to buttons

* fix: format

* fix: icon minwidth

---------
