+++ apps/web/screens/Grants/Grant/Grant.css.ts
@@ -1,0 +1,7 @@
+import { globalStyle, style } from '@vanilla-extract/css'
+
+export const noUnderline = style({})
+
+globalStyle(`${noUnderline} span`, {
+  boxShadow: 'none',
+})
